Refactored exponent setting; variants -> submodules
